Simplify the traits, Endian adapter, add tests for Hash
Hash no longer has the byte order as the parameter which made it unusable.

EndianInput provides a ByteOrder it works with as an associated type.
Its methods are no longer parameterized with a ByteOrder.

An Endian adapter struct is provided to wrap digest types and provide
an implementation of EndianInput with the selected byte order.
Type aliases BigEndian and LittleEndian specialize the byte order to
provide an ergonomic construction API.

Exercise the primitive type impls on Hash with unit tests.